Turn CSV of addresses into a PDF of Avery 8160 mailing labels
Prompt: I have a CSV that contains an address book for our friends and family - it has columns for name, address, city, state, zip, and country. I'd like to print out the addresses as mailing labels in Avery 8160 format. Can you write a python script that will take the CSV and turn it into a PDF that I can print onto the Avery 8160 mailing labels?
Signed-off-by: Shai Sachs <shaisachs@gmail.com>